refactor(review): share review diff helpers
