+++ src/review-checkpoints.ts
@@ -1,31 +1,16 @@
 import { mkdtemp, rm } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import { parsePatchFiles } from "@pierre/diffs";
 import { git, getGitEligibility, safeWorkspaceRefSegment } from "./git.js";
+import {
+  parseReviewFiles,
+  summarizeReviewFiles,
+  type ReviewChangesResult,
+} from "./review-diff.js";
+
+export type { ReviewChangesResult, ReviewFile, ReviewSummary } from "./review-diff.js";
 
 export type ReviewSince = "last_shown" | "workspace_open";
-
-export interface ReviewSummary {
-  files: number;
-  additions: number;
-  removals: number;
-}
-
-export interface ReviewFile {
-  path: string;
-  previousPath?: string;
-  type: "change" | "rename-pure" | "rename-changed" | "new" | "deleted";
-  additions: number;
-  removals: number;
-}
-
-export interface ReviewChangesResult {
-  result: string;
-  summary: ReviewSummary;
-  files: ReviewFile[];
-  patch: string;
-}
 
 interface WorkspaceReviewState {
   root: string;
@@ -124,7 +109,7 @@ export function createReviewCheckpointManager(): ReviewCheckpointManager {
         maxBuffer: REVIEW_DIFF_MAX_BUFFER,
       })).stdout;
       const files = parseReviewFiles(patch);
-      const summary = summarizeFiles(files);
+      const summary = summarizeReviewFiles(files);
 
       if (markReviewed) {
         await git(state.gitRoot, ["update-ref", state.baselineRef, current]);
@@ -146,47 +131,6 @@ export function createReviewCheckpointManager(): ReviewCheckpointManager {
       };
     },
   };
-}
-
-function parseReviewFiles(patch: string): ReviewFile[] {
-  if (patch.length === 0) return [];
-
-  try {
-    return parsePatchFiles(patch, "review", true).flatMap((parsedPatch) =>
-      parsedPatch.files.map((file) => {
-        const stats = file.hunks.reduce(
-          (total, hunk) => ({
-            additions: total.additions + hunk.additionLines,
-            removals: total.removals + hunk.deletionLines,
-          }),
-          { additions: 0, removals: 0 },
-        );
-
-        return {
-          path: file.name,
-          previousPath: file.prevName,
-          type: reviewFileType(file.type),
-          ...stats,
-        };
-      }),
-    );
-  } catch (error) {
-    const detail = error instanceof Error ? error.message : String(error);
-    throw new Error(`Review diff could not be rendered: ${detail}`);
-  }
-}
-
-function reviewFileType(type: string): ReviewFile["type"] {
-  switch (type) {
-    case "rename-pure":
-    case "rename-changed":
-    case "new":
-    case "deleted":
-    case "change":
-      return type;
-    default:
-      return "change";
-  }
 }
 
 function assertWorkspaceRoot(
@@ -290,15 +234,4 @@ function checkpointEnv(indexPath: string): NodeJS.ProcessEnv {
     GIT_COMMITTER_NAME: "DevSpace",
     GIT_COMMITTER_EMAIL: "devspace@users.noreply.local",
   };
-}
-
-function summarizeFiles(files: ReviewFile[]): ReviewSummary {
-  return files.reduce<ReviewSummary>(
-    (summary, file) => ({
-      files: summary.files + 1,
-      additions: summary.additions + file.additions,
-      removals: summary.removals + file.removals,
-    }),
-    { files: 0, additions: 0, removals: 0 },
-  );
 }
